Dead code moved
